refactor: cartItemId 자료형 수정
- int -> Long

+++ src/main/java/com/example/shopping_mall/cart/cart_item/CartItem.java
@@ -21,7 +21,7 @@ public class CartItem {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int cartItemId;
+  private Long cartItemId;
 
   @ManyToOne
   private Cart cart;
